# docs/swagger.yaml
basePath: /api/
definitions:
  api.ParamSearch:
    properties:
      filter:
        additionalProperties: true
        type: object
      keyword:
        additionalProperties:
          type: string
        type: object
      limit:
        type: integer
      skip:
        type: integer
      sort:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  api.ReplyData-int64:
    properties:
      data:
        type: integer
      error:
        type: string
    type: object
...
  api.ReplyData-types_Client:
    properties:
      data:
        $ref: '#/definitions/types.Client'
      error:
        type: string
    type: object
  api.ReplyData-types_Device:
    properties:
      data:
        $ref: '#/definitions/types.Device'
      error:
        type: string
    type: object
...
  api.ReplyList-types_Client:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Client'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  api.ReplyList-types_Device:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Device'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
...
  types.Client:
    properties:
      addr:
        description: 地址，主机名或IP
        type: string
      created:
        description: 创建时间
        type: string
      desc:
        description: 说明
        type: string
      disabled:
        type: boolean
      heartbeat:
        description: 心跳包
        type: string
      id:
        description: ID
        type: string
      name:
        description: 名称
        type: string
      net:
        description: 类型 tcp udp
        type: string
      poller_interval:
        description: 采集间隔
        type: integer
      poller_period:
        description: 采集周期
        type: integer
      port:
        description: 端口号
        type: integer
      protocol_name:
        description: 协议 rtu tcp parallel-tcp
        type: string
      protocol_options:
        description: 协议参数
        type: string
      retry_maximum:
        description: 最大次数
        type: integer
      retry_timeout:
        description: 重试时间
        type: integer
    type: object
...
  types.Device:
    properties:
      created:
        type: string
      desc:
        description: 说明
        type: string
      disabled:
        type: boolean
      id:
        type: string
      name:
        description: 名称
        type: string
      product_id:
        description: 产品ID
        type: string
      slave:
        description: 从站号
        type: integer
      tunnel_id:
        description: 通道
        type: string
    type: object
...
info:
  contact: {}
  description: API文档
  title: 物联大师网关接口文档
  version: 1.0 版本
paths:
  /client/{id}:
    get:
      consumes:
      - application/json
      description: 获取客户端
      parameters:
      - description: 客户端ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Client'
      summary: 获取客户端
      tags:
      - client
    post:
      consumes:
      - application/json
      description: 修改客户端
      parameters:
      - description: 客户端ID
        in: path
        name: id
        required: true
        type: integer
      - description: 客户端信息
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/types.Client'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Client'
      summary: 修改客户端
      tags:
      - client
  /client/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除客户端
      parameters:
      - description: 客户端ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Client'
      summary: 删除客户端
      tags:
      - client
  /client/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用客户端
      parameters:
      - description: 客户端ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Client'
      summary: 禁用客户端
      tags:
      - client
  /client/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用客户端
      parameters:
      - description: 客户端ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Client'
      summary: 启用客户端
      tags:
      - client
  /client/count:
    post:
      consumes:
      - application/json
      description: 查询客户端数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 查询客户端数量
      tags:
      - client
  /client/create:
    post:
      consumes:
      - application/json
      description: 创建客户端
      parameters:
      - description: 客户端信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Client'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Client'
      summary: 创建客户端
      tags:
      - client
...
  /client/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入客户端
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 导入客户端
      tags:
      - product
  /client/list:
    get:
      consumes:
      - application/json
      description: 查询客户端
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Client'
      summary: 查询客户端
      tags:
      - client
  /client/search:
    post:
      consumes:
      - application/json
      description: 查询客户端
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Client'
      summary: 查询客户端
      tags:
      - client
...
  /device/{id}:
    get:
      consumes:
      - application/json
      description: 获取设备
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Device'
      summary: 获取设备
      tags:
      - device
    post:
      consumes:
      - application/json
      description: 修改设备
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      - description: 设备信息
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/types.Device'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Device'
      summary: 修改设备
      tags:
      - device
  /device/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除设备
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Device'
      summary: 删除设备
      tags:
      - device
  /device/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用设备
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Device'
      summary: 禁用设备
      tags:
      - device
  /device/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用设备
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Device'
      summary: 启用设备
      tags:
      - device
  /device/count:
    post:
      consumes:
      - application/json
      description: 查询设备数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 查询设备数量
      tags:
      - device
  /device/create:
    post:
      consumes:
      - application/json
      description: 创建设备
      parameters:
      - description: 设备信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Device'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Device'
      summary: 创建设备
      tags:
      - device
...
  /device/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入设备
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 导入设备
      tags:
      - product
  /device/list:
    get:
      consumes:
      - application/json
      description: 查询设备
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Device'
      summary: 查询设备
      tags:
      - device
  /device/search:
    post:
      consumes:
      - application/json
      description: 查询设备
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Device'
      summary: 查询设备
      tags:
      - device
...
swagger: "2.0"